Fixed assert.

+++ gip/transformation/StraightLineHoughTransformation.cpp
@@ -73,7 +73,7 @@ namespace gip {
           DestinationImage::Pixel* destRow = dest;
           for (int theta = 0; trigo < endOfTrigo; ++trigo, ++theta, destRow += width) { // vote for lines
             int rho = static_cast<int>(x * trigo->cosine + y * trigo->sine + halfWidth);
-            ASSERT(rho < width);
+            ASSERT(static_cast<unsigned int>(rho) < width);
             ASSERT(rho >= 0);
             ++destRow[rho]; // TAG: alternatively use gradient as weight
           }
